# swagger.yml
openapi: 3.0.1
info:
  title: Restaurants Service API
  description: Restaurants from GoOutSafe with Microservices
  version: "0.1"
servers:
- url: /
paths:
...
  /restaurants/id/{owner_email}:
    get:
      tags:
      - Restaurants
      summary: Get the id of a restaurant
      operationId: app.get_restaurant_id_by_owner_email
      parameters:
      - name: owner_email
        in: path
        description: Email of the owner of the restaurant
        required: true
        schema:
          type: string
      responses:
        200:
          description: Return the restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantId'
        404:
          description: Owner not found
          content: 
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}:
    get:
      tags:
      - Restaurants
      summary: Get a single restaurant
      operationId: app.get_restaurant
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        200:
          description: Return restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        404:
          description: Restaurant not found
          content: 
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Restaurants
      summary: Delete a restaurant with all its info
      operationId: app.delete_restaurant
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        200:
          description: "The operation is done"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        400:
          description: "The code returned if there is some error on server, e.g: param not specified"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Restaurant not found
          content: 
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}/name:
    get:
      tags:
      - Restaurants
      summary: Get the name of the restaurant
      operationId: app.get_restaurant_name
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        200:
          description: Return restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        404:
          description: Restaurant not found
          content: 
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}/info:
    get:
      tags:
      - Restaurants
      summary: Get more info about the restaurant
      operationId: app.get_restaurant_more_info
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        200:
          description: Return restaurant
          content:
            application/json:
              schema:
                type: object
                properties:
                  Restaurant_info:
                    type: object
                    $ref: '#/components/schemas/Info'
        404:
          description: Restaurant not found
          content: 
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/search/{name}:
    get:
      tags:
      - Restaurants
      summary: Get a restaurant restaurant with keyword name
      operationId: app.get_restaurants_by_keyword
      parameters:
      - name: name
        in: path
        description: name to search
        required: true
        schema:
          type: string
      responses:
        200:
          description: Return all restaurants with keyword name
          content:
            application/json:
              schema:
                type: object
                properties:
                  restaurants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
  /restaurants/{restaurant_id}/menu:
    get:
      tags:
      - Menu
      summary: Get the menu of a single restaurant
      operationId: app.get_menus
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        200:
          description: Return the menus of the restaurant
          content:
            application/json:
              schema:
                type: object
                properties:
                  menus:
                    type: array
                    items:
                      $ref: '#/components/schemas/Menus'
        404:
          description: Restaurant not found
          content: 
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/menu/{menu_id}:
    post:
      tags:
      - Menu
      summary: Create a new photo for the menu
      operationId: app.create_menu_photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoCreation"
      parameters:
      - name: menu_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        201:
          description: "Photo added"
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/MenuPhotos'
        404:
          description: Error restaurant not found
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
        409:
          description: Error URL already used
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}/dishes:
    get:
      tags:
      - Dishes
      summary: Get the dishes of a single restaurant
      operationId: app.get_dishes
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        200:
          description: Return the dishes of the restaurant
          content:
            application/json:
              schema:
                type: object
                properties:
                  dishes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dishes'
        404:
          description: Restaurant not found
          content: 
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
    post:
      tags:
      - Dishes
      summary: Create a new dish in the restaurant
      operationId: app.create_dish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DishCreation"
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        201:
          description: "Dish added"
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Dishes'
        404:
          description: Error restaurant not found
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}/openings:
    get:
      tags:
      - Opening Hours
      summary: Get the opening hours of a single restaurant
      operationId: app.get_openings
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        200:
          description: Return the opening hours of the restaurant
          content:
            application/json:
              schema:
                type: object
                properties:
                  openings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Openings'
        404:
          description: Restaurant not found
          content: 
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/table/{table_id}:
    get:
      tags:
        - Tables
      summary: Get the table with all restaurant info
      operationId: app.get_table
      parameters:
        - name: table_id
          in: path
          description: Table's Unique identifier
          required: true
          schema:
            pattern: ^[a-zA-Z0-9-]+$
            type: string
      responses:
        200:
          description: Return the restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotalTable'
        404:
          description: Owner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}/tables:
    get:
      tags:
      - Tables
      summary: Get the tables of a single restaurant
      operationId: app.get_tables
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        200:
          description: Return the tables of the restaurant
          content:
            application/json:
              schema:
                type: object
                properties:
                  tables:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tables'
        404:
          description: Restaurant not found
          content: 
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
    post:
      tags:
      - Tables
      summary: Create a table in the restaurant
      operationId: app.create_table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableCreation"
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        201:
          description: "Table added to restaurant"
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Tables'
        404:
          description: Generic error
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}/photos:
    get:
      tags:
      - Photos
      summary: Get the photos of a single restaurant
      operationId: app.get_photos
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        200:
          description: Return the photos of the restaurant
          content:
            application/json:
              schema:
                type: object
                properties:
                  photos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Photos'
        404:
          description: Restaurant not found
          content: 
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
    post:
      tags:
      - Photos
      summary: Create a new photo for the restaurant
      operationId: app.create_photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoCreation"
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        201:
          description: "Photo added"
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Photos'
        404:
          description: Error restaurant not found
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
        409:
          description: Error URL already used
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}/reviews:
    get:
      tags:
      - Reviews
      summary: Get the reviews of a single restaurant
      operationId: app.get_reviews
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        200:
          description: Return the reviews of the restaurant
          content:
            application/json:
              schema:
                type: object
                properties:
                  Reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reviews'
        404:
          description: Restaurant not found
          content: 
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
    post:
      tags:
      - Reviews
      summary: Create a new review for the restaurant
      operationId: app.create_review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreation"
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      responses:
        201:
          description: "Review added"
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Reviews'
        404:
          description: Error restaurant not found
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}/reviews/{number}:
    get:
      tags:
      - Reviews
      summary: Get the a specified number of random reviews
      operationId: app.get_random_reviews
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          pattern: ^[a-zA-Z0-9-]+$
          type: string
      - name: number
        in: path
        description: number of reviews to get
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Return the random reviews of the restaurant
          content:
            application/json:
              schema:
                type: object
                properties:
                  Reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reviews'
        404:
          description: Restaurant not found
          content: 
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}/avg_rating:
    get:
      tags:
      - Reviews
      summary: Get avg rating of a restaurant
      operationId: app.get_avg_rating_restaurant
      parameters:
      - name: restaurant_id
        in: path
        description: Restaurant's Unique identifier
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Return the avg reviews of the restaurant
          content:
            application/json:
              schema:
                type: number
        500:
          description: Generic error
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/calculate_rating_for_all_restaurant:
      get:
        tags:
        - Reviews
        summary: Calculate rating for all restaurant (celery)
        operationId: app.calculate_rating_for_all_restaurant
        responses:
          200:
            description: "The operation is done"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Result'
          500:
            description: Generic error
            content:
              application/json:
                schema:
                    $ref: '#/components/schemas/Error'
  /restaurants/create:
    post:
      tags:
      - Restaurants
      summary: Create a restaurant
      operationId: app.create_restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantModel"
      responses:
        201:
          description: "Restaurant is been created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantId"
        409:
          description: Generic error
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurants/update:
    put:
      tags:
      - Restaurants
      summary: Update a restaurant
      operationId: app.update_restaurant_info
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantId"
      responses:
        200:
          description: "Restaurant is been updated"
        500:
          description: Generic error
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'
  /restaurant/menu/{dish_id}:
    delete:
      tags:
        - Dishes
      summary: Delete a dish
      operationId: app.delete_dish
      parameters:
        - in: path
          name: dish_id
          schema:
            type: integer
          required: true
          description: The dish_id ID
      responses:
        200:
          description: "The operation is done"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        400:
          description: "The code returned if there is some error on server, e.g: param not specified"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
...
components:
  schemas:
    Error:
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    Result:
      required:
        - result
      type: object
      properties:
        result:
          type: string
          default: Type of result
          example: "Message to describe what happen on server"
...
    RestaurantId:
      required:
        - id
        - name
        - lat
        - lon
        - phone
        - covid_measures
        - avg_time
        - rating
        - owner_email
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          example: 123
        name:
          maxLength: 100
          minLength: 1
          type: string
          description: Restaurant's name
          example: "My Pizza Restaurant"
        lat:
          type: number
          description: Restaurant Latitude
          example: 42.42
        lon:
          type: number
          description: Restaurant Longitude
          example: 42.42
        phone:
          type: integer
          minLength: 5
          maxLength: 50
          description: Restaurant Phone Number
          example: 050123456
        covid_measures:
          type: string
          maxLength: 500
          description: Covid measures adopted in restaurant
          example: "Distance between tables"
        avg_time:
          type: integer
          description: Restaurant avg time
          example: 42
        rating:
          type: number
          description: Restaurant Rating
          example: 0
        owner_email:
          type: string
          description: Restaurant owner's email
          example: "mario.rossi@mail.it"
...